requirements.txt
pycryptodome
pyqt5
pyperclip